Revise README and refactor webhook faber&alice
